+++ cpp/src/sonar/Sonar.cpp
@@ -13,9 +13,9 @@ int Sonar::getEchoPin() {
     return echoPin;
 }
 
-float Sonar::calculateDistance(float rawVal) {
-    const float vs = 331.45 + 0.62*20;
-    float t = rawVal / 1000.0 / 1000.0 / 2;
-    float d = t*vs;
+double Sonar::calculateDistance(double rawVal) {
+    const double vs = 331.45 + 0.62*20;
+    double t = rawVal / 1000.0 / 1000.0 / 2;
+    double d = t*vs;
     return d;
 }
